+++ .gitignore
@@ -6,3 +6,4 @@
 *.webp binary
 *.woff binary
 *.woff2 binary
+node_modules
