+++ public/index.html
@@ -34,7 +34,7 @@
     <!-- Контент -->
   </main>
 
-  <script src="../src/main.js"></script>
+  <script src="./bundle.js"></script>
 </body>
 
 </html>
