+++ src/components/board.js
@@ -1,0 +1,15 @@
+const createBoardTemplate = () => {
+  return (
+    `<section class="board container">
+      <div class="board__filter-list">
+        <a href="#" class="board__filter">SORT BY DEFAULT</a>
+        <a href="#" class="board__filter">SORT BY DATE up</a>
+        <a href="#" class="board__filter">SORT BY DATE down</a>
+      </div>
+
+      <div class="board__tasks"></div>
+    </section>`
+  );
+};
+
+export default createBoardTemplate;
